Build our own mediawiki and use it

+++ build_scripts/build_wikibase_docker.sh
@@ -1,7 +1,12 @@
 #!/bin/bash
 set -e
 
-cp "$TARBALL_PATH" Docker/DockerWikibase
-docker build --pull --build-arg MEDIAWIKI_IMAGE_NAME=$MEDIAWIKI_IMAGE_NAME Docker/DockerWikibase/ -t "$1"
+docker load -i "artifacts/$MEDIAWIKI_IMAGE_NAME.docker.tar.gz"
 
-docker save "$1" | gzip -9f > "$(pwd)"/artifacts/"$1".docker.tar.gz
+if [ -f "$TARBALL_PATH" ]; then
+    cp "$TARBALL_PATH" Docker/DockerWikibase
+fi
+
+docker build --build-arg MEDIAWIKI_IMAGE_NAME=$MEDIAWIKI_IMAGE_NAME Docker/DockerWikibase/ -t "$1"
+
+docker save "$1" | gzip -9f > artifacts/"$1".docker.tar.gz
